diode-suzieq-agent/pyproject.toml
[project]
name = "diode-suzieq-agent"
version = "0.0.1"  # Overwritten during the build process
description = "NetBox Labs, Diode SuzieQ Agent"
readme = "README.md" 
requires-python = ">3.8.1, < 3.10"
license = {file = "LICENSE.txt"}
authors = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" } # Optional
]
maintainers = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" } # Optional
]

classifiers = [  # Optional
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]

dependencies = [
    "suzieq==0.22.0",
    "netboxlabs-diode-sdk==0.0.1",
]

[project.optional-dependencies] # Optional
dev = ["black", "check-manifest", "ruff"]
test = ["coverage", "pytest", "pytest-cov"]

[project.urls]  # Optional
"Homepage" = "https://netboxlabs.com/"

[project.scripts]  # Optional

[tool.setuptools]
packages = [
    "diode_suzieq.cli.cli",
    "diode_suzieq.extension",
    "diode_suzieq.client",
    "diode_suzieq.parser",
]

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.ruff]
line-length = 140
exclude = [
    "netboxlabs/diode/sdk/diode/*",
    "netboxlabs/diode/sdk/validate/*",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["C", "D", "E", "F", "I", "R", "UP", "W"]
ignore = ["F401", "D203", "D212", "D400", "D401", "D404", "RET504"]